Dev Testing updates
 - documentation
 - split up tests to apialerts and config
 - update readme
 - remove excess debug logging
